+++ lib/src/generator/renderers/context_renderer.dart
@@ -1,8 +1,10 @@
 import '../../models/analytics_event.dart';
 import '../../util/string_utils.dart';
 import '../../util/type_mapper.dart';
+import 'base_renderer.dart';
 
-class ContextRenderer {
+class ContextRenderer extends BaseRenderer {
+  const ContextRenderer();
   String renderContextFile(
     String contextName,
     List<AnalyticsParameter> properties,
@@ -11,11 +13,8 @@ class ContextRenderer {
     final camelContextName = StringUtils.toCamelCase(contextName);
 
     final buffer = StringBuffer();
-    buffer.writeln('// GENERATED CODE - DO NOT MODIFY BY HAND');
-    buffer.writeln('// ignore_for_file: type=lint, unused_import');
-    buffer.writeln();
-    buffer.writeln("import 'package:analytics_gen/analytics_gen.dart';");
-    buffer.writeln();
+    buffer.write(renderFileHeader(includeCoverageIgnore: false));
+    buffer.write(renderImports(['package:analytics_gen/analytics_gen.dart']));
 
     buffer.writeln('/// Capability interface for $pascalName');
     buffer.writeln(
@@ -52,12 +51,10 @@ class ContextRenderer {
       if (prop.allowedValues != null && prop.allowedValues!.isNotEmpty) {
         final constName =
             'allowed${StringUtils.capitalizePascal(camelName)}Values';
-        final encodedValues = prop.allowedValues!
-            .map((value) => '\'${StringUtils.escapeSingleQuoted(value)}\'')
-            .join(', ');
-        final joinedValues = prop.allowedValues!.join(', ');
+        final encodedValues = encodeAllowedValues(prop.allowedValues!);
+        final joinedValues = joinAllowedValues(prop.allowedValues!);
 
-        buffer.write(_renderAllowedValuesCheck(
+        buffer.write(renderAllowedValuesCheck(
           camelParam: 'value',
           constName: constName,
           encodedValues: encodedValues,
@@ -74,30 +71,6 @@ class ContextRenderer {
 
     buffer.writeln('}');
 
-    return buffer.toString();
-  }
-
-  String _renderAllowedValuesCheck({
-    required String camelParam,
-    required String constName,
-    required String encodedValues,
-    required String joinedValues,
-    required bool isNullable,
-  }) {
-    final buffer = StringBuffer();
-    buffer.writeln('    const $constName = <String>{$encodedValues};');
-
-    final condition = isNullable
-        ? 'if ($camelParam != null && !$constName.contains($camelParam)) {'
-        : 'if (!$constName.contains($camelParam)) {';
-
-    buffer.writeln('    $condition');
-    buffer.writeln('      throw ArgumentError.value(');
-    buffer.writeln('        $camelParam,');
-    buffer.writeln("        '$camelParam',");
-    buffer.writeln("        'must be one of $joinedValues',");
-    buffer.writeln('      );');
-    buffer.writeln('    }');
     return buffer.toString();
   }
 }